Changed movement to be a float and not an int

+++ src/main/java/Engine/Collisions/AABB.java
@@ -47,6 +47,6 @@ public class AABB {
     }
 
     // Setters
-    public void setX(int x) { this.x = x; }
-    public void setY(int y) { this.y = y; }
+    public void setX(float x) { this.x = x; }
+    public void setY(float y) { this.y = y; }
 }
